Longitudinal Gain Scheduled vs Midpoint Comparison
Comparison between gain scheduled controller vs midpoint as a function of Va and altitude. Will do this same comparison with theta and Va as scheduling parameters to determine if this is a better way forward (because legal limit is ~= 150 m altitude for a consumer drone). If dynamics vary substantially with theta then will go this route because it's more appropriate in practice. If not, will need to note that altitude and Va are a good academic exercise for a generic flight platform (not constrained by legalities).