Selected Va & Alpha as Scheduling Variables
Selected Va & Alpha as scheduling variables because altitude didn't make sense for legal reasons, theta didn't make sense for practical reasons (ie. not directly measured) and Va makes sense because it has an obvious impact and it's directly measurable. Alpha is directly measurable and has a large impact but it's difficult to determine how useful it will be in practice without knowing the kinds of alpha variations expected - this will be reviewed at a later stage and alpha might be dropped and just Va used if the variation is small.